django/requirements.txt
Django==5.2.6
django-ninja==1.4.3
pydantic==2.11.9
psycopg2-binary==2.9.9
redis==6.4.0
celery==5.5.3
django-celery-beat==2.8.1
django-celery-results==2.6.0
gunicorn==23.0.0
watchdog==6.0.0
django-model-utils==5.0.0
PyJWT==2.8.0
flower==2.0.1
django-cors-headers==4.9.0
openai==2.1.0
requests==2.32.5
django-storages==1.14.6
boto3
Pillow